% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Atlantis_Catch_to_MSSPM.R
\name{convertAtlantisCatchToMSSPM}
\alias{convertAtlantisCatchToMSSPM}
\title{Atlantis to MSSPM Catch Conversion}
\usage{
convertAtlantisCatchToMSSPM(inputDataFrame, startYear, outputFile)
}
\arguments{
\item{inputDataFrame}{The Atlantis input simulated catch data frame}

\item{startYear}{The year to start the conversion (earlier years will be skipped)}

\item{outputfile}{The output csv data file containing the MSSPM catch formatted data}
}
\value{
n/a
}
\description{
Conversion of Atlantis Simulated Catch Index file to MSSPM Catch csv file
}
\examples{
inputDataFrame <- mskeyrun::simCatchIndex
startYear <- 55
outputFile <- '/home/rklasky/test/testCatch.csv'

convertAtlantisCatchToMSSPM(inputDataFrame,startYear,outputFile)

}
