% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Atlantis_ObsBM_to_MSSPM.R
\name{convertAtlantisObsBMtoMSSPM}
\alias{convertAtlantisObsBMtoMSSPM}
\title{Atlantis to MSSPM Biomass Conversion}
\usage{
convertAtlantisObsBMtoMSSPM(
  speciesMappingDataFrame,
  observedBMDataFrame,
  surveySeason,
  outputFile
)
}
\arguments{
\item{speciesMappingDataFrame}{The Atlantis data frame that contains a mapping of the SVSPP numeric values to corresponding species names (i.e., modelName)}

\item{observedBMDataFrame}{The Atlantis input survey index (i.e. biomass) data frame}

\item{surveySeason}{The survey season data to convert (Atlantis file currently has FALL and SPRING)}

\item{outputfile}{The output csv data file containing the MSSPM observed biomass formatted data}
}
\value{
n/a
}
\description{
Conversion of Atlantis Survey Index file to MSSPM Observed Biomass csv file
}
\examples{

speciesMappingDataFrame <- mskeyrun::focalSpecies # needed to map SVSPP value to modelName
observedBMDataFrame <- mskeyrun::surveyIndexA4
surveySeason <- 'SPRING' # Use 'FALL' or "SPRING"
outputFile <- paste('/home/rklasky/test/testObsBM_',surveySeason,'.csv',sep="")

convertAtlantisObsBMtoMSSPM(speciesMappingDataFrame,observedBMDataFrame,surveySeason,outputFile)

}
